requirements.txt
apache-airflow-providers-google==10.3.0
soda-core-bigquery==3.5.4
soda-core-scientific==3.5.4
google-cloud-bigquery-storage==2.26.0
google-cloud-bigquery==3.18.0
google-cloud-storage==2.18.2
astronomer-cosmos==1.10.1
dbt-bigquery==1.8.2
protobuf==4.25.2
